requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv~=1.0.0